+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 All notable changes to this project will be documented in this file. 
 
+
+> 17 March 2020
+- Added a `site:url` command for adding more URLs to an existing site. Also updated the `site:list` command, that now
+will output more information.
+#### [0.1.0-beta8](https://github.com/digitalist-se/extratools/compare/0.1.0-beta7...0.1.0-beta8)
+
 > 17 March 2020
 - Added support for Matomo default env. variables, https://hub.docker.com/_/matomo?tab=description
 #### [0.1.0-beta7](https://github.com/digitalist-se/extratools/compare/0.1.0-beta6...0.1.0-beta7)
-
 
 > 17 March 2020
 - Added site:delete command